requirements.txt
httpx==0.27.0
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11
PyYAML==6.0.2
cloudscraper==1.2.71
python-jobspy
